--- src/components/common/TabNavigation.tsx
@@ -1,27 +1,0 @@
-import React from 'react';
-import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import Home from '../../screens/home';
-import MyPage from '../../screens/mypage';
-import Community from '../../screens/community';
-
-const Tab = createBottomTabNavigator();
-
-const TabNavigation = () => {
-  return (
-    <Tab.Navigator>
-      <Tab.Screen name="Home" component={Home} options={{headerShown: false}} />
-      <Tab.Screen
-        name="Community"
-        component={Community}
-        options={{headerShown: false}}
-      />
-      <Tab.Screen
-        name="MyPage"
-        component={MyPage}
-        options={{headerShown: false}}
-      />
-    </Tab.Navigator>
-  );
-};
-
-export default TabNavigation;
